Fix true-colour truncation: item-glow rainbow + faded status bars
Two more colour-truncation sites the Phase 6c sweep missed, both from
storing/passing a 32-bit ARGB Get16BPPColor result through a 16-bit
path so PixFromColor16 re-decoded the low bits as an RGB565 token.

- vobject_blitters.cpp outline blitters: the item-glow outline colour
  (a full PIXEL parameter) was cast (UINT16) before PixFromColor16, so
  the grey/white/red/yellow glow ramp came out cycling through the
  whole rainbow. Pass the full PIXEL; PixFromColor16 already passes an
  already-expanded ARGB value through unchanged. Fixed in all 7 outline
  variants (base/Clip/Z/ZClip/ZNB/ZPixelateObscured[Clip]).

- Interface Utils.cpp: usLineColor (life/bandage/bleeding/breath/morale
  bars) was declared UINT16 but assigned Get16BPPColor and passed to
  RectangleDraw(...PIXEL...), desaturating the pure R/G/B bars. Widened
  to PIXEL.

These are framebuffer-content bugs independent of the SDL3 presentation
path; they reproduce regardless of window scaling.

+++ sgp/vobject_blitters.cpp
@@ -6575,7 +6575,13 @@ BOOLEAN Blt8BPPDataTo16BPPBufferOutline( PIXEL *pBuffer, UINT32 uiDestPitchBYTES
 					const UINT8 v = *src++;
 					if (v == 254) {
 						if (fDoOutline) {
-							*rowDest = PixFromColor16((UINT16)s16BPPColor);
+							// Pass the full PIXEL: s16BPPColor is already a
+							// screen-format colour (true-colour ARGB8888 at
+							// 32bpp). Truncating it to UINT16 here dropped the
+							// high half, so PixFromColor16 decoded the low bits
+							// as an RGB565 token -- which turned the grey
+							// item-glow ramp into a rainbow.
+							*rowDest = PixFromColor16(s16BPPColor);
 						}
 					} else {
 						*rowDest = p16BPPPalette[v];
@@ -6686,7 +6692,7 @@ BOOLEAN Blt8BPPDataTo16BPPBufferOutlineClip( PIXEL *pBuffer, UINT32 uiDestPitchB
 						const INT32 dx = srcX - LeftSkip;
 						if (v == 254) {
 							if (fDoOutline) {
-								rowDest[dx] = PixFromColor16((UINT16)s16BPPColor);
+								rowDest[dx] = PixFromColor16(s16BPPColor);
 							}
 						} else {
 							rowDest[dx] = p16BPPPalette[v];
@@ -6809,7 +6815,7 @@ BOOLEAN Blt8BPPDataTo16BPPBufferOutlineZClip( PIXEL *pBuffer, UINT32 uiDestPitch
 						if (usZValue >= rowZ[dx]) {
 							if (v == 254) {
 								if (fDoOutline) {
-									rowDest[dx] = PixFromColor16((UINT16)s16BPPColor);
+									rowDest[dx] = PixFromColor16(s16BPPColor);
 								}
 							} else {
 								rowZ[dx]    = usZValue;
@@ -6943,7 +6949,7 @@ BOOLEAN Blt8BPPDataTo16BPPBufferOutlineZPixelateObscuredClip( PIXEL *pBuffer, UI
 					}
 					if (v == 254) {
 						if (fDoOutline) {
-							rowDest[dx] = PixFromColor16((UINT16)s16BPPColor);
+							rowDest[dx] = PixFromColor16(s16BPPColor);
 						}
 					} else {
 						rowDest[dx] = p16BPPPalette[v];
@@ -7217,7 +7223,7 @@ BOOLEAN Blt8BPPDataTo16BPPBufferOutlineZ( PIXEL *pBuffer, UINT32 uiDestPitchBYTE
 					if (usZValue >= *rowZ) {
 						if (v == 254) {
 							if (fDoOutline) {
-								*rowDest = PixFromColor16((UINT16)s16BPPColor);
+								*rowDest = PixFromColor16(s16BPPColor);
 							}
 						} else {
 							*rowZ    = usZValue;
@@ -7315,7 +7321,7 @@ BOOLEAN Blt8BPPDataTo16BPPBufferOutlineZPixelateObscured( PIXEL *pBuffer, UINT32
 					if (render) {
 						if (v == 254) {
 							if (fDoOutline) {
-								*rowDest = PixFromColor16((UINT16)s16BPPColor);
+								*rowDest = PixFromColor16(s16BPPColor);
 							}
 						} else {
 							*rowDest = p16BPPPalette[v];
@@ -7398,7 +7404,7 @@ BOOLEAN Blt8BPPDataTo16BPPBufferOutlineZNB( PIXEL *pBuffer, UINT32 uiDestPitchBY
 					if (usZValue >= *rowZ) {
 						if (v == 254) {
 							if (fDoOutline) {
-								*rowDest = PixFromColor16((UINT16)s16BPPColor);
+								*rowDest = PixFromColor16(s16BPPColor);
 							}
 						} else {
 							*rowDest = p16BPPPalette[v];
